backend/requirements.txt
asgiref==3.9.1
Django==5.2.5
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.1
pip-autoremove==0.10.0
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
setuptools==80.9.0
sqlparse==0.5.3
typing_extensions==4.14.1
whitenoise==6.9.0
